Floating point width fix
As offsetWidth always brings values rounded up if they have floating points, sometimes the resizing could break the text in an additional line with only one word just before the ellipsis line. To fix it the method getBoundClientRect is used as it brings width and height from the element as it is (with the floating point).